implement message truncation feature
If the message cannot be packed into a packet
with a size limit of 512 byte the algorithm
tries to remove the additional and authoriy section.
If this is sufficient the truncation bit is set and
the message is transmitted.

The algorithm should be more fine-grained and also consider
to split the particular sections to fit into to packet limit.

Signed-off-by: Hagen Paul Pfeifer <hagen@jauu.net>