implement EDNS0 support
This commit implement complete EDNS0 support.
The daemon is variable per configuration file via
edns0-mode and edns0-size.

edns0-mode is per default on. If the resolver request
edns0 we provide edns0. If edns0-mode is set to off we
sent a packet with max 512 byte payload and do not set
this RR.

Via edns0-size the local limit can be adjusted. Per default
the limit is set to 8k. But it can be increased via

edns0-size 16384

to, for example, 16k.

The next step is to implement edns0 toward the server side. This is
currently not implemented but this is not complicated because the
infrastructure is already implemented.

Signed-off-by: Hagen Paul Pfeifer <hagen@jauu.net>