fix edns0 bug to provide edns0 also in a absence of other sections
Until this fix the packet sent back to the resolver
was send without a edns0 type in the additional section
if no other additional section was provided by the forwarder.

With this patch we always generate a edns0 type secion if
the client inform us that he also support this feature.

Signed-off-by: Hagen Paul Pfeifer <hagen@jauu.net>